% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khaiiir-package.R
\docType{package}
\name{khaiiir-package}
\alias{khaiiir}
\alias{khaiiir-package}
\title{khaiiir: R Binding for khaiii which is Korean Analyzer}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (\href{https://orcid.org/0000-0001-6474-2570}{ORCID})

}
\keyword{internal}
